+++ example/5a-list-buckets.js
