+++ example/5b-add-remove-bucket.js
