+++ example/6d-delete-file-from-bucket.js
